feat: distinguish source and target branches on PR builds
When a build is triggered by a Pull Request the `branch` property would contains either the branch targeted by the PR or the branch the Pull Request originated from, depending on the CI service.
This change harmonize that behavior and now, for Pull Request builds, set the following properties:
- `branch`: the branch target by the Pull Request (e.g. `master`)
- `prBranch`: the branch the Pull Request originated from (e.g. `feature/my-feature`)

BREAKING CHANGE: The value of `branch` will change for certain CI services when building a Pull Request (and only in that case).

For the following services, `branch` will contains the Pull Request target rather than the branch the Pull Request originated from: Bitrise, Buildkite, Cirrus, Drone, Jenkins, Shippable, VSTS.

For the following services, `branch` will be `undefined` as no variable is provided to determine the Pull Request target: Buddy, CircleCI, Sail, Semaphore.

For the following services, `branch` will not change and still contains the Pull Request target: Travis, Appveyor.

The `prBranch` property can be used to determine the branch the Pull Request originated from.