fix: bump minimal node version from `10.13` to `10.17` to remain compatible with dependencies
This is technically a breaking change, but I do not know which dependency requires 10.17 and do not have the time to investigate further. Plus the alternative to pinning said dependencie might open up `env-ci` to known security vulnerability which we do not want. It is a tradeoff, and all things considered we decided to release it as a fix